# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation Presence",
        "version": "1.0.0"
    },
    "paths": {
        "/api/attendances": {
            "get": {
                "tags": [
                    "attendance"
                ],
                "summary": "Get a list of attendances",
                "operationId": "f7564e4f8fdec50c27cf2a372050327d",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by user name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/attendances/{id}": {
            "delete": {
                "tags": [
                    "attendance"
                ],
                "summary": "Delete an attendance record",
                "operationId": "3b1350863e6c4643d660bad18a27d91e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the attendance record to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendance deleted successfully"
                    },
                    "404": {
                        "description": "Attendance not found"
                    }
                }
            }
        },
        "/api/companies": {
            "get": {
                "tags": [
                    "company"
                ],
                "summary": "List all companies",
                "operationId": "54f966c056050f9db4fff6ee5a45a76a",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter companies by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            },
            "post": {
                "tags": [
                    "company"
                ],
                "summary": "Create a new company",
                "operationId": "00e2db7b0f71c32b377f4dfc4dd685a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the company",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email of the company",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "address": {
                                        "description": "Address of the company",
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "description": "Latitude of the company's location",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "description": "Longitude of the company's location",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "radius_km": {
                                        "description": "Radius in kilometers around the company's location",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "time_in": {
                                        "description": "Time when the company opens",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "time_out": {
                                        "description": "Time when the company closes",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Company created successfully"
                    }
                }
            }
        },
        "/api/companies/{id}": {
            "put": {
                "tags": [
                    "company"
                ],
                "summary": "Update a company",
                "operationId": "6a1d7a4a2b9cbdfe7749cef78e6cec57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the company",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email of the company",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "address": {
                                        "description": "Address of the company",
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "description": "Latitude of the company's location",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "description": "Longitude of the company's location",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "radius_km": {
                                        "description": "Radius in kilometers around the company's location",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "time_in": {
                                        "description": "Time when the company opens",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "time_out": {
                                        "description": "Time when the company closes",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Company updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "company"
                ],
                "summary": "Delete a company",
                "operationId": "49d767f608bffc95ad39368bf1766693",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Company deleted successfully"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                }
            }
        },
        "/api/permissions": {
            "get": {
                "tags": [
                    "permission"
                ],
                "summary": "List all permissions",
                "operationId": "27bb1606a7b293d1d49f57bb94ae71f7",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter permissions by user name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/permissions/{id}": {
            "get": {
                "tags": [
                    "permission"
                ],
                "summary": "Get a specific permission",
                "operationId": "397983813683c0c602fc1a036992367a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "404": {
                        "description": "Permission not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "permission"
                ],
                "summary": "Update a specific permission",
                "operationId": "db11d4d00da3ece5b99187df6f1e86e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "Status of the permission",
                                        "type": "string",
                                        "example": "approved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission updated successfully"
                    },
                    "404": {
                        "description": "Permission not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "permission"
                ],
                "summary": "Delete a specific permission",
                "operationId": "f0afbd22579524a5bb83b653f7dd18dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Permission deleted successfully"
                    },
                    "404": {
                        "description": "Permission not found"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "List all users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter users by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            },
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Store a newly created user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "employee_type": {
                                        "type": "string",
                                        "example": "full-time"
                                    },
                                    "department": {
                                        "type": "string",
                                        "example": "IT"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Developer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Update a specific user",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Delete a specific user",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "attendance",
            "description": "attendance"
        },
        {
            "name": "company",
            "description": "company"
        },
        {
            "name": "permission",
            "description": "permission"
        },
        {
            "name": "user",
            "description": "user"
        }
    ]
}